Remove casts and enum all sprites as ImageIndex
In order to remove excessive casting, kSpriteIdNull was rectified
to properly populate as the max uint32_t value (required limits
std library in ImageIndexType.h).

Any casts of kSpriteIdNull to uint32_t were removed (since
ImageIndex is an alias of that type already). In
EditorParkEntrace.cpp, kSpriteIdNull is no longer casted to
ImageIndex, since it is declared as ImageIndex type.

Sprite values in sprites.h now enumerate as type ImageIndex.

+++ src/openrct2-ui/interface/Widget.cpp
@@ -823,7 +823,7 @@ namespace OpenRCT2::Ui
         const auto& widget = w.widgets[widgetIndex];
 
         // Get the image
-        if (static_cast<int32_t>(widget.image.ToUInt32()) == kSpriteIdNull)
+        if (widget.image.ToUInt32() == kSpriteIdNull)
             return;
         auto image = widget.image;
